CircuitProcessorの機能をCircuitへ移動
今更ながら何でCircuitProcessorと別にしたのか謎
