--- Assets/Scripts/porideki/circuit/CircuitProcessor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e5f3646d19db4644a9390675cd06b2da
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
